+++ certifisafe-back/service/certificate_service.go
@@ -21,9 +21,9 @@ var (
 )
 
 type ICertificateService interface {
-	UpdateCertificate(id int32, certificate model.Certificate) (model.Certificate, error)
-	GetCertificate(id int32) (model.Certificate, error)
-	DeleteCertificate(id int32) error
+	UpdateCertificate(id int64, certificate model.Certificate) (model.Certificate, error)
+	GetCertificate(id int64) (model.Certificate, error)
+	DeleteCertificate(id int64) error
 	CreateCertificate(certificate model.Certificate) (model.Certificate, error)
 }
 
@@ -37,7 +37,7 @@ func NewDefaultCertificateService(cRepo repository.ICertificateRepository) *Defa
 	}
 }
 
-func (d *DefaultCertificateService) UpdateCertificate(id int32, certificate model.Certificate) (model.Certificate, error) {
+func (d *DefaultCertificateService) UpdateCertificate(id int64, certificate model.Certificate) (model.Certificate, error) {
 	// if id <= 0 {
 	// 	return ErrIDIsNotValid
 	// }
@@ -53,11 +53,11 @@ func (d *DefaultCertificateService) UpdateCertificate(id int32, certificate mode
 
 	return model.Certificate{}, nil
 }
-func (d *DefaultCertificateService) GetCertificate(id int32) (model.Certificate, error) {
+func (d *DefaultCertificateService) GetCertificate(id int64) (model.Certificate, error) {
 	certificate, err := d.certificateRepo.GetCertificate(id)
 	return certificate, err
 }
-func (d *DefaultCertificateService) DeleteCertificate(id int32) error {
+func (d *DefaultCertificateService) DeleteCertificate(id int64) error {
 
 	return nil
 }
